fix(macos): extract ContentView into its own file for PreviewHost
MingaApp.swift contained ContentView, StartupOverlay, PaneHeightKey,
TitleBarDragRegion, and TitleBarDragNSView alongside the @main app entry.
PreviewHost excludes MingaApp.swift and needs these types in a separate
file. Move them to ContentView.swift so both targets compile.
